backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
httpx>=0.26.0
python-dotenv==1.0.1
python-jose[cryptography]>=3.3.0
supabase>=2.9.0
websockets>=13,<16
boto3>=1.28.0

# Async task queue
celery[redis]==5.3.6
redis==5.0.1

# LLM report generation and email delivery
google-generativeai>=0.8.0
resend>=2.0.0
